requirements.txt
pycryptodome==3.14.1
cryptography==37.0.2
